+++ CHANGES.md
@@ -5,8 +5,9 @@
 ## v0.0.0 - 2024-xx-xx
 - Implement `linksmith inventory` and `linksmith output-formats`
   subcommands, based on `sphobjinv` and others. Thanks, @bskinn.
-- Implement `linksmith anansi suggest`, also available as `anansi`,
+- Anansi: Implement `linksmith anansi suggest`, also available as `anansi`,
   to easily suggest terms of a few curated community projects.
   Thanks, @bskinn.
 - Accept `linksmith inventory` without `INFILES` argument, implementing
   auto-discovery of `objects.inv` in local current working directory.
+- Anansi: Manage project list in YAML file `curated.yaml`, not Python.